+++ lib/Db/MemberMapper.php
@@ -20,6 +20,7 @@
  */
 namespace OCA\Majordomo\Db;
 
+use OC\DB\QueryBuilder\Parameter;
 use OCP\DB\QueryBuilder\IQueryBuilder;
 use OCP\IDBConnection;
 
@@ -52,6 +53,37 @@ class MemberMapper extends \OCP\AppFramework\Db\QBMapper {
                 $qb->expr()->eq("list_id", $qb->createNamedParameter($id, IQueryBuilder::PARAM_INT)),
                 $qb->expr()->in("type", $qb->createNamedParameter($types, IQueryBuilder::PARAM_STR_ARRAY)),
             ));
+    }
+
+    /**
+     * Rather complex query to find all lists visible to a user.
+     *
+     * @param $user string the user ID
+     * @param $groups string[] the group IDs the user is a member of
+     * @return Member[] the member entries matching the user
+     * @throws \OCP\DB\Exception
+     */
+    public function findAllByUserAndGroups($user, $groups, $listId) {
+        $qb = $this->db->getQueryBuilder();
+        $or = $qb->expr()->orX();
+        if ($listId !== NULL) {
+            $or->add($qb->expr()->eq("list_id", $qb->createNamedParameter($listId, IQueryBuilder::PARAM_INT)));
+        }
+        if (!empty($user)) {
+            $or->add($qb->expr()->andX(
+                $qb->expr()->in("type", $qb->createNamedParameter(Member::TYPES_USER, IQueryBuilder::PARAM_STR_ARRAY)),
+                $qb->expr()->eq("reference", $qb->createNamedParameter($user))
+            ));
+        }
+        if (!empty($group)) {
+            $or->add($qb->expr()->andX(
+                $qb->expr()->in("type", $qb->createNamedParameter(Member::TYPES_GROUP, IQueryBuilder::PARAM_STR_ARRAY)),
+                $qb->expr()->eq("reference", $qb->createNamedParameter($groups, IQueryBuilder::PARAM_STR_ARRAY))
+            ));
+        }
+        return $this->findEntities($qb->select("*")
+            ->from("majordomo_members")
+            ->where($or));
     }
 
 }