+++ lib/Service/WebhookSettings.php
@@ -1,7 +1,8 @@
+<?php
 /**
- * @copyright Copyright (c) 2020 Marco Ziech <marco+nc@ziech.net>
+ * @copyright Copyright (c) 2024 Marco Ziech <marco+nc@ziech.net>
  *
- * @license AGPL-3.0-or-later
+ * @license GNU AGPL version 3 or any later version
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -17,21 +18,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
+namespace OCA\Majordomo\Service;
 
-import { generateUrl } from "@nextcloud/router";
-import axios from "@nextcloud/axios";
-
-const BASEPATH = '/apps/majordomo/api';
-export default new class Api {
-    path(path) {
-        return generateUrl(BASEPATH + path);
-    }
-
-    get(path) {
-        return axios.get(generateUrl(BASEPATH + path)).then(response => response.data);
-    }
-
-    post(path, data) {
-        return axios.post(generateUrl(BASEPATH + path), data).then(response => response.data);
-    }
-};
+class WebhookSettings {
+    public $enabled;
+    public $token;
+}